requirements.txt
numpy>=1.26.0
PyYAML>=6.0.1
pypdf>=3.17.4
tqdm>=4.66.1
openai>=1.12.0
transformers>=4.39.3
sentence-transformers>=2.6.1
torch>=2.1.2
matplotlib>=3.8.0
networkx>=3.2.1
pdfplumber>=0.11.0
pyvis>=0.3.2
